back : error handle to validate userId, productId and quantity as number

+++ server/src/modules/cart/cartActions.ts
@@ -1,15 +1,22 @@
 import type { RequestHandler } from "express";
+import { StatusCodes } from "http-status-codes";
 import cartRepository from "./cartRepository";
 
 const browse: RequestHandler = async (req, res, next) => {
   try {
     const userId = Number(req.params.userId);
+    if (Number.isNaN(userId)) {
+      res
+        .status(StatusCodes.BAD_REQUEST)
+        .json({ error: "Identifiant doit être un nombre" });
+      return;
+    }
     const products = await cartRepository.find(userId);
 
     if (products == null) {
-      res.sendStatus(404);
+      res.status(StatusCodes.NOT_FOUND);
     } else {
-      res.json(products);
+      res.status(StatusCodes.OK).json(products);
     }
   } catch (err) {
     next(err);
@@ -24,13 +31,24 @@ const update: RequestHandler = async (req, res, next) => {
       quantity: req.body.quantity,
     };
 
+    if (
+      Number.isNaN(updatedCart.userId) ||
+      Number.isNaN(updatedCart.productId) ||
+      Number.isNaN(updatedCart.quantity)
+    ) {
+      res
+        .status(StatusCodes.BAD_REQUEST)
+        .json({ error: "Paramétres invalides" });
+      return;
+    }
+
     await cartRepository.updateCartQuantity(
       updatedCart.userId,
       updatedCart.productId,
       updatedCart.quantity,
     );
 
-    res.json(updatedCart);
+    res.status(StatusCodes.OK).json(updatedCart);
   } catch (err) {
     next(err);
   }
@@ -43,9 +61,19 @@ const remove: RequestHandler = async (req, res, next) => {
       userId: Number(req.params.userId),
     };
 
+    if (
+      Number.isNaN(deletedCart.userId) ||
+      Number.isNaN(deletedCart.productId)
+    ) {
+      res
+        .status(StatusCodes.BAD_REQUEST)
+        .json({ error: "Paramétres invalides" });
+      return;
+    }
+
     await cartRepository.delete(deletedCart.userId, deletedCart.productId);
 
-    res.json(deletedCart);
+    res.status(StatusCodes.OK).json(deletedCart);
   } catch (err) {
     next(err);
   }
